ddl: move the backfill scheduler to a separate file
